food_model/gptgram_model/requirements.txt
recipe-scrapers
ingredient-parser-nlp
inflect
